refactor(system): 将岗位 ID 类型从 Integer 改为 Long
- 修改了 SysPostController、SysPostService、SysPostServiceImpl 和 SysPostVo 中的岗位 ID 类型
- 移除了 SysPostVo 中的 createTime、updateTime、createBy 和 updateBy 字段- 为 SysPostVo 添加了 @EqualsAndHashCode 注解，使其继承自 BaseVo

+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/service/impl/SysPostServiceImpl.java
@@ -67,7 +67,7 @@ public class SysPostServiceImpl extends ServiceImpl<SysPostMapper, SysPost>
      * @return 操作结果
      */
     @Override
-    public boolean deletePost(List<Integer> ids) {
+    public boolean deletePost(List<Long> ids) {
         return removeByIds(ids);
     }
 
@@ -78,7 +78,7 @@ public class SysPostServiceImpl extends ServiceImpl<SysPostMapper, SysPost>
      * @return 操作结果
      */
     @Override
-    public SysPost getPostById(Integer id) {
+    public SysPost getPostById(Long id) {
         return getById(id);
     }
